remove MR scripts; fix KernelDensityEstimator
Former-commit-id: 6e38bc0cb70496a0cca2f183d73e25706bd62727

+++ datadriven/src/sgpp/datadriven/application/KernelDensityEstimator.cpp
@@ -569,7 +569,7 @@ void SilvermansRule::optimizeBandwidths(KernelDensityEstimator* kde, base::DataV
       }
     }
 
-    if (numBorder > static_cast<double>(numSamples) / 20.) {
+    if (static_cast<double>(numBorder) > static_cast<double>(numSamples) / 20.) {
       flag[idim] = 0.5;
     }
